refact: Transactions page to enable SSR
